Prepare locations for migration to serialization
